Downsampling of map is working, improved goal selection, added config params. RAM usage < 100 MB on a 200 x 200 meter map.
git-svn-id: https://rarepos.cs.uni-tuebingen.de/svn-path/rabot/trunk/ros/navigation@3338 690284e3-2886-4bbc-99f2-a987503dca67

+++ a_star/include/a_star/Node.h
@@ -35,8 +35,8 @@ public:
 
 private:
 	waypoint_t mPosition;
-	double mDistance;
-	double mPrediction;
+    float mDistance;
+    float mPrediction;
 	Node* mParent;
 	bool mVisited;
 	bool mInitialized;
